Lab11/Storage/requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
clickclick==20.10.2
colorama==0.4.5
connexion==2.14.1
Flask
greenlet==1.1.3
idna==3.4
importlib-metadata
inflection==0.5.1
itsdangerous
Jinja2
jsonschema
kazoo==2.5.0
MarkupSafe
mysql-connector-python==8.0.30
packaging==21.3
pykafka==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent
PyYAML==6.0
requests
six==1.16.0
SQLAlchemy==1.4.41
swagger-ui-bundle==0.0.9
tabulate
urllib3==1.26.12
Werkzeug
zipp
